+++ CHANGELOG.md
@@ -5,11 +5,11 @@
 ### New features
 
 * Indent and font-lock forms that start with `let-`, `while-` or `when-` like their counterparts.
-* Namespace font-locking according to clojure.lang.LispReader
 
 ### Bugs fixed
 
 * Namespaces can now use the full palette of legal symbol characters.
+* Namespace font-locking according to clojure.lang.LispReader.
 
 ## 5.0.1 (15/11/2015)